+++ mg_custom_nodes/VrchStudio_comfyui-web-viewer_20260121/__init__.py
@@ -10,7 +10,7 @@ from .nodes.logic_nodes import *
 from .nodes.midi_nodes import *
 from .nodes.audio_music2emo_node import *
 
-__version__ = "1.1.11"
+__version__ = "1.1.12"
 
 NODE_CLASS_MAPPINGS = {
     "VrchAnyOSCControlNode": VrchAnyOSCControlNode,